+++ lib/index.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const mapnik = require('@carto/mapnik');
-const timeoutDecorator = require('./utils/timeout-decorator');
+const timeoutDecorator = require('../utils/timeout-decorator');
 const createMapPool = require('./map-pool');
 const normalizeURI = require('./uri');
