+++ lib/map-pool.js
