+++ lib/uri.js
