+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "@carto/tilelive-bridge",
   "version": "2.5.1-cdb12",
-  "main": "./index.js",
+  "main": "lib/index.js",
   "description": "Datasource => vector tiles bridge backend for tilelive",
   "repository": {
     "type": "git",
@@ -23,7 +23,7 @@
     "tape": "^4.9.1"
   },
   "scripts": {
-    "lint": "eslint index.js",
+    "lint": "eslint lib/",
     "pretest": "npm run lint",
     "test": "tape test/*.js"
   }
